requirements.txt
Django==3.0.2
pyserial==3.4
